backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy
pandas
scikit-learn
xgboost
transformers
torch
ollama
presidio-analyzer
presidio-anonymizer
pdfplumber
python-docx
tldextract
librosa
python-dotenv
openai-whisper
ffmpeg-python
easyocr
pillow
requests
pymongo
passlib[bcrypt]
python-jose[cryptography]